refactor(debug): reduce duplication in debugger tests
